+++ ui/src/mixins/index.js
@@ -1,7 +1,7 @@
 import contextMixin from './context.js'
 import formMixin from './form.js'
 import generatorMixin from './generator.js'
-import lazyLoadingFilterMixin from './lazy-loading-filter.js'
+import searchFilterMixin from './search-filter.js'
 import passwordMixin from './password.js'
 import viewMixin from './view.js'
 
@@ -9,7 +9,7 @@ export {
   contextMixin,
   formMixin,
   generatorMixin,
-  lazyLoadingFilterMixin,
+  searchFilterMixin,
   passwordMixin,
   viewMixin
 }